Move zarr helper functions to utils for unit testability
Relocate _resolve_analyzer_dir and _strip_non_numeric_properties from
spike_sorting.py (which activates a schema on import) to a pure
utils.spike_sorting_utils module so they can be imported and unit tested
without a database connection. No behavior change.
