+++ aeon/dj_pipeline/utils/spike_sorting_utils.py
@@ -1,0 +1,26 @@
+"""Pure helpers for the spike sorting pipeline's zarr/binary intermediate handling.
+
+These functions have no database or table dependencies, so they live here (rather
+than in spike_sorting.py) to keep them importable and unit-testable without
+activating the spike_sorting schema.
+"""
+
+from pathlib import Path
+
+import numpy as np
+
+
+def _resolve_analyzer_dir(output_dir: Path) -> Path:
+    """Find sorting analyzer directory, checking both binary and zarr paths."""
+    analyzer_dir = output_dir / "sorting_analyzer"
+    if not analyzer_dir.exists():
+        analyzer_dir = output_dir / "sorting_analyzer.zarr"
+    return analyzer_dir
+
+
+def _strip_non_numeric_properties(si_recording) -> None:
+    """Remove non-numeric recording properties that zarr v2 cannot serialize."""
+    for prop in list(si_recording.get_property_keys()):
+        values = si_recording.get_property(prop)
+        if np.asarray(values).dtype.kind not in ("f", "i", "u", "b"):
+            si_recording.delete_property(prop)
